backend/requirements.txt
psutil>=5.8.0
mohawk>=1.1.0
requests>=2.25.1
urllib3>=1.26.4
GitPython>=3.1.14
gunicorn>=20.1.0
flask>=1.0.2
Flask-Cors>=3.0.10
regex>=2021.4.4